coi-data-cli-PI3-Release/requirements.txt
obspy
requests
requests_mock
pandas
msgpack-python
pyyaml
termcolor